td-shim-enroll-key: refactor to share code with td-shim::verify
Refactor the td-shim-enroll-key to prepare sharing code with
td-shim::verify.
