+++ td-shim-enroll-key/src/main.rs
@@ -1,103 +1,132 @@
 // Copyright (c) 2021 Intel Corporation
+// Copyright (c) 2022 Alibaba Cloud
 //
 // SPDX-License-Identifier: BSD-2-Clause-Patent
 
 #[macro_use]
 extern crate clap;
 
-use std::ptr::slice_from_raw_parts;
+use std::convert::TryFrom;
+use std::path::PathBuf;
 use std::str::FromStr;
 use std::vec::Vec;
-use std::{convert::TryFrom, mem::size_of};
 use std::{env, io, path::Path};
 
 use log::{error, trace, LevelFilter};
-
-use r_efi::efi::Guid;
 use ring::digest;
-use scroll::{Pread, Pwrite};
-use td_layout::build_time::{TD_SHIM_CONFIG_OFFSET, TD_SHIM_CONFIG_SIZE, TD_SHIM_FIRMWARE_SIZE};
-use td_shim_ld::{write_u24, FvFfsHeader, FvHeader, InputData, OutputFile};
-use uefi_pi::fv::*;
-
-mod public_key;
-use self::public_key::{
+use td_layout::build_time::{TD_SHIM_CONFIG_OFFSET, TD_SHIM_FIRMWARE_SIZE};
+use td_shim_enroll_key::public_key::{
     RsaPublicKeyInfo, SubjectPublicKeyInfo, ID_EC_PUBKEY_OID, RSA_PUBKEY_OID, SECP384R1_OID,
 };
+use td_shim_enroll_key::{
+    build_cfv_ffs_header, build_cfv_header, CfvPubKeyFileHeader, CFV_FILE_HEADER_PUBKEY_GUID,
+    PUBKEY_FILE_STRUCT_VERSION_V1, PUBKEY_HASH_ALGORITHM_SHA384,
+};
+use td_shim_ld::linker::{InputData, OutputFile};
 
-const CFV_FFS_HEADER_GUID: Guid = Guid::from_fields(
-    0x77a2742e,
-    0x9340,
-    0x4ac9,
-    0x8f,
-    0x85,
-    &[0xb7, 0xb9, 0x78, 0x58, 0x0, 0x21],
-); // {77A2742E-9340-4AC9-8F85-B7B978580021}
-
-const FS_DATA_HEADER_GUID: Guid = Guid::from_fields(
-    0xbe8f65a3,
-    0xa83b,
-    0x415c,
-    0xa1,
-    0xfb,
-    &[0xf7, 0x8e, 0x10, 0x5e, 0x82, 0x4e],
-); // {BE8F65A3-A83B-415C-A1FB-F78E105E824E}
-
-const PUBKEY_FILE_STRUCT_VERSION: u32 = 0x01;
 const TDSHIM_SB_NAME: &str = "final.sb.bin";
 
-#[repr(C, align(4))]
-#[derive(Pread, Pwrite)]
-struct CfvDataFileHeader {
-    pub type_guid: [u8; 16],
-    pub struct_version: u32,
-    pub length: u32,
-    pub hash_algorithm: u64,
-    pub reserved: u32,
-}
+fn enroll_key(
+    tdshim_file: &str,
+    key_file: &str,
+    output_file: PathBuf,
+    hash_alg: &'static digest::Algorithm,
+) -> io::Result<()> {
+    let tdshim_bin = InputData::new(
+        tdshim_file,
+        TD_SHIM_FIRMWARE_SIZE as usize..=TD_SHIM_FIRMWARE_SIZE as usize,
+        "shim binary",
+    )?;
+    let key_data = InputData::new(key_file, 1..=1024 * 1024, "public key")?;
+    let key = SubjectPublicKeyInfo::try_from(key_data.as_bytes()).map_err(|e| {
+        error!("Can not load key from file {}: {}", key_file, e);
+        io::Error::new(io::ErrorKind::Other, "invalid key data")
+    })?;
 
-impl CfvDataFileHeader {
-    pub fn as_bytes(&self) -> &[u8] {
-        unsafe { &*slice_from_raw_parts(self as *const Self as *const u8, size_of::<Self>()) }
+    let mut public_bytes: Vec<u8> = Vec::new();
+    match key.algorithm.algorithm {
+        ID_EC_PUBKEY_OID => {
+            if let Some(curve) = key.algorithm.parameters {
+                if curve.as_bytes() != SECP384R1_OID.as_bytes() {
+                    error!("Unsupported Named Curve from file {}", key_file);
+                    return Err(io::Error::new(
+                        io::ErrorKind::Other,
+                        "unsupported Named Curve",
+                    ));
+                }
+                if key.subject_public_key.as_bytes()[0] != 0x04 {
+                    error!("Invalid SECP384R1 public key from file {}", key_file);
+                    return Err(io::Error::new(
+                        io::ErrorKind::Other,
+                        "Invalid SECP384R1 public key",
+                    ));
+                }
+                public_bytes.extend_from_slice(&key.subject_public_key.as_bytes()[1..]);
+            } else {
+                error!("Invalid algorithm parameter from file {}", key_file);
+                return Err(io::Error::new(
+                    io::ErrorKind::Other,
+                    "Invalid key algorithm parameter",
+                ));
+            }
+        }
+        RSA_PUBKEY_OID => {
+            let pubkey =
+                RsaPublicKeyInfo::try_from(key.subject_public_key.as_bytes()).map_err(|e| {
+                    error!("Invalid key from file {}: {}", key_file, e);
+                    io::Error::new(io::ErrorKind::Other, "invalid key from file")
+                })?;
+            public_bytes.extend_from_slice(pubkey.modulus.as_bytes());
+            let mut exp_bytes = [0u8; 8];
+            if pubkey.exponents.as_bytes().len() > 8 {
+                error!("Invalid exponent size from key file {}", key_file);
+                return Err(io::Error::new(
+                    io::ErrorKind::Other,
+                    "Invalid exponent size",
+                ));
+            }
+            exp_bytes[8 - pubkey.exponents.as_bytes().len()..]
+                .copy_from_slice(pubkey.exponents.as_bytes());
+            public_bytes.extend_from_slice(&exp_bytes);
+        }
+        t => {
+            error!("Unsupported key type {} from file {}", t, key_file);
+            return Err(io::Error::new(io::ErrorKind::Other, "unsupported key type"));
+        }
     }
+
+    let hash = digest::digest(hash_alg, public_bytes.as_slice());
+    let hash = hash.as_ref();
+
+    //Build public key header in CFV
+    let pub_key_header = CfvPubKeyFileHeader {
+        type_guid: *CFV_FILE_HEADER_PUBKEY_GUID.as_bytes(),
+        struct_version: PUBKEY_FILE_STRUCT_VERSION_V1,
+        length: (36 + hash.len()) as u32,
+        hash_algorithm: PUBKEY_HASH_ALGORITHM_SHA384,
+        reserved: 0,
+    };
+
+    // Create and write the td-shim binary with key enrolled.
+    let mut output = OutputFile::new(output_file)?;
+    let cfv_header = build_cfv_header();
+    let cfv_ffs_header = build_cfv_ffs_header();
+
+    output.seek_and_write(0, tdshim_bin.as_bytes(), "enrolled shim binary")?;
+    output.seek_and_write(
+        TD_SHIM_CONFIG_OFFSET as u64,
+        cfv_header.as_bytes(),
+        "firmware volume header",
+    )?;
+    output.write(cfv_ffs_header.as_bytes(), "firmware volume fs header")?;
+    output.write(pub_key_header.as_bytes(), "firmware key")?;
+    output.write(hash, "firmware hash value")?;
+    output.flush()?;
+
+    Ok(())
 }
 
-fn build_cfv_header() -> FvHeader {
-    let mut cfv_header = FvHeader::default();
-
-    cfv_header
-        .fv_header
-        .file_system_guid
-        .copy_from_slice(FIRMWARE_FILE_SYSTEM3_GUID.as_bytes());
-    cfv_header.fv_header.fv_length = TD_SHIM_CONFIG_SIZE as u64;
-    cfv_header.fv_header.checksum = 0xdc0a;
-    cfv_header.fv_block_map[0].num_blocks = (TD_SHIM_CONFIG_SIZE as u32) / 0x1000;
-    cfv_header.fv_block_map[0].length = 0x1000;
-    cfv_header.fv_ext_header.ext_header_size = 0x14;
-
-    cfv_header
-}
-
-fn build_cfv_ffs_header() -> FvFfsHeader {
-    let mut cfv_ffs_header = FvFfsHeader::default();
-    cfv_ffs_header
-        .ffs_header
-        .name
-        .copy_from_slice(CFV_FFS_HEADER_GUID.as_bytes());
-
-    cfv_ffs_header.ffs_header.integrity_check = 0xaa4c;
-    cfv_ffs_header.ffs_header.r#type = FV_FILETYPE_RAW;
-    cfv_ffs_header.ffs_header.attributes = 0x00;
-    write_u24(
-        TD_SHIM_CONFIG_SIZE - size_of::<FvHeader>() as u32,
-        &mut cfv_ffs_header.ffs_header.size,
-    );
-    cfv_ffs_header.ffs_header.state = 0x07u8;
-
-    cfv_ffs_header
-}
-
-fn main() -> std::io::Result<()> {
+fn main() -> io::Result<()> {
     use env_logger::Env;
     let env = Env::default()
         .filter_or("MY_LOG_LEVEL", "info")
@@ -173,96 +202,5 @@ fn main() -> std::io::Result<()> {
         }
     };
 
-    let tdshim_bin = InputData::new(
-        tdshim_file,
-        TD_SHIM_FIRMWARE_SIZE as usize..=TD_SHIM_FIRMWARE_SIZE as usize,
-        "shim binary",
-    )?;
-    let key_data = InputData::new(key_file, 1..=1024 * 1024, "private key")?;
-    let key = SubjectPublicKeyInfo::try_from(key_data.as_bytes()).map_err(|e| {
-        error!("Can not load key from file {}: {}", key_file, e);
-        io::Error::new(io::ErrorKind::Other, "invalid key data")
-    })?;
-
-    let mut public_bytes: Vec<u8> = Vec::new();
-    match key.algorithm.algorithm {
-        ID_EC_PUBKEY_OID => {
-            if let Some(curve) = key.algorithm.parameters {
-                if curve.as_bytes() != SECP384R1_OID.as_bytes() {
-                    error!("Unsupported Named Curve from file {}", key_file);
-                    return Err(io::Error::new(
-                        io::ErrorKind::Other,
-                        "unsupported Named Curve",
-                    ));
-                }
-                if key.subject_public_key.as_bytes()[0] != 0x04 {
-                    error!("Invalid SECP384R1 public key from file {}", key_file);
-                    return Err(io::Error::new(
-                        io::ErrorKind::Other,
-                        "Invalid SECP384R1 public key",
-                    ));
-                }
-                public_bytes.extend_from_slice(&key.subject_public_key.as_bytes()[1..]);
-            } else {
-                error!("Invalid algorithm parameter from file {}", key_file);
-                return Err(io::Error::new(
-                    io::ErrorKind::Other,
-                    "Invalid key algorithm parameter",
-                ));
-            }
-        }
-        RSA_PUBKEY_OID => {
-            let pubkey =
-                RsaPublicKeyInfo::try_from(key.subject_public_key.as_bytes()).map_err(|e| {
-                    error!("Invalid key from file {}: {}", key_file, e);
-                    io::Error::new(io::ErrorKind::Other, "invalid key from file")
-                })?;
-            public_bytes.extend_from_slice(pubkey.modulus.as_bytes());
-            let mut exp_bytes = [0u8; 8];
-            if pubkey.exponents.as_bytes().len() > 8 {
-                error!("Invalid exponent size from key file {}", key_file);
-                return Err(io::Error::new(
-                    io::ErrorKind::Other,
-                    "Invalid exponent size",
-                ));
-            }
-            exp_bytes[8 - pubkey.exponents.as_bytes().len()..]
-                .copy_from_slice(pubkey.exponents.as_bytes());
-            public_bytes.extend_from_slice(&exp_bytes);
-        }
-        t => {
-            error!("Unsupported key type {} from file {}", t, key_file);
-            return Err(io::Error::new(io::ErrorKind::Other, "unsupported key type"));
-        }
-    }
-
-    let hash = digest::digest(hash_alg, public_bytes.as_slice());
-    let hash = hash.as_ref();
-
-    //Build public key header in CFV
-    let pub_key_header = CfvDataFileHeader {
-        type_guid: *FS_DATA_HEADER_GUID.as_bytes(),
-        struct_version: PUBKEY_FILE_STRUCT_VERSION,
-        length: (36 + hash.len()) as u32,
-        hash_algorithm: 1,
-        reserved: 0,
-    };
-
-    // Create and write the td-shim binary with key enrolled.
-    let mut output = OutputFile::new(output_file)?;
-    let cfv_header = build_cfv_header();
-    let cfv_ffs_header = build_cfv_ffs_header();
-
-    output.seek_and_write(0, tdshim_bin.as_bytes(), "enrolled shim binary")?;
-    output.seek_and_write(
-        TD_SHIM_CONFIG_OFFSET as u64,
-        cfv_header.as_bytes(),
-        "firmware volume header",
-    )?;
-    output.write(cfv_ffs_header.as_bytes(), "firmware volume fs header")?;
-    output.write(pub_key_header.as_bytes(), "firmware key")?;
-    output.write(hash, "firmware hash value")?;
-    output.flush()?;
-
-    Ok(())
+    enroll_key(tdshim_file, key_file, output_file, hash_alg)
 }
